requirements.txt
langchain-community>=0.2.0
pymupdf>=1.24.0
python-docx>=1.1.0
openai>=1.0.0
python-dotenv>=1.0.1
